% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpsnip-package.R
\docType{package}
\name{perpsnip-package}
\alias{perpsnip-package}
\alias{class-package}
\title{perpsnip: 'tidymodels' Interface for 'PerpetualBooster'}
\description{
Provides a 'parsnip' model specification for the 'PerpetualBooster' algorithm implemented in the 'perpetual' package. PerpetualBooster is a self-generalizing gradient boosting machine that eliminates the need for hyperparameter search by automatically tuning itself within a user-supplied computational budget. This package exposes classification, regression, and (optionally) censored regression modes so that PerpetualBooster can be used inside any 'tidymodels' workflow, including hyperparameter tuning with 'tune' and performance evaluation with 'yardstick'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshuamarie/perpsnip}
  \item Report bugs at \url{https://github.com/joshuamarie/perpsnip/issues}
}

}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{package}
